+++ src/__tests__/helpers/sqlMock.ts
@@ -1,0 +1,12 @@
+import * as s from '../../sql'
+
+// mock functions that get data from postgres and give them the right types
+type UnPromisify<F> = F extends (...args: infer A) => Promise<infer T> ? (...args: A) => T : never
+export const getPluginRows = (s.getPluginRows as unknown) as jest.MockedFunction<UnPromisify<typeof s.getPluginRows>>
+export const getPluginAttachmentRows = (s.getPluginAttachmentRows as unknown) as jest.MockedFunction<
+    UnPromisify<typeof s.getPluginAttachmentRows>
+>
+export const getPluginConfigRows = (s.getPluginConfigRows as unknown) as jest.MockedFunction<
+    UnPromisify<typeof s.getPluginConfigRows>
+>
+export const setError = (s.setError as unknown) as jest.MockedFunction<UnPromisify<typeof s.setError>>
