+++ src/__tests__/plugins.test.ts
@@ -1,129 +1,16 @@
 import { runPlugins, setupPlugins } from '../plugins'
 import { createServer } from '../server'
-import { Plugin, PluginAttachmentDB, PluginConfig, PluginsServer } from '../types'
-import * as s from '../sql'
-import AdmZip from 'adm-zip'
+import { PluginsServer } from '../types'
 import { PluginEvent } from 'posthog-plugins/src/types'
-import * as fs from 'fs'
-import * as os from 'os'
-import * as path from 'path'
-
-// mock functions that get data from postgres and give them the right types
+import {
+    mockPluginTempFolder,
+    mockPluginWithArchive,
+    plugin60,
+    pluginAttachment1,
+    pluginConfig39,
+} from './helpers/plugins'
+import { getPluginAttachmentRows, getPluginConfigRows, getPluginRows, setError } from './helpers/sqlMock'
 jest.mock('../sql')
-type UnPromisify<F> = F extends (...args: infer A) => Promise<infer T> ? (...args: A) => T : never
-const getPluginRows = (s.getPluginRows as unknown) as jest.MockedFunction<UnPromisify<typeof s.getPluginRows>>
-const getPluginAttachmentRows = (s.getPluginAttachmentRows as unknown) as jest.MockedFunction<
-    UnPromisify<typeof s.getPluginAttachmentRows>
->
-const getPluginConfigRows = (s.getPluginConfigRows as unknown) as jest.MockedFunction<
-    UnPromisify<typeof s.getPluginConfigRows>
->
-const setError = (s.setError as unknown) as jest.MockedFunction<UnPromisify<typeof s.setError>>
-
-function createZipBuffer(name: string, { indexJs, pluginJson }: { indexJs?: string; pluginJson?: string }): Buffer {
-    const zip = new AdmZip()
-    if (indexJs) {
-        zip.addFile('testplugin/index.js', Buffer.alloc(indexJs.length, indexJs))
-    }
-    if (pluginJson) {
-        zip.addFile('testplugin/plugin.json', Buffer.alloc(pluginJson.length, pluginJson))
-    } else {
-        zip.addFile(
-            'testplugin/plugin.json',
-            new Buffer(
-                JSON.stringify({
-                    name,
-                    description: 'just for testing',
-                    url: 'http://example.com/plugin',
-                    config: {},
-                    main: 'index.js',
-                })
-            )
-        )
-    }
-    return zip.toBuffer()
-}
-
-const mockPluginWithArchive = (indexJs: string, pluginJson?: string) => ({
-    ...plugin60,
-    archive: createZipBuffer('posthog-maxmind-plugin', { indexJs, pluginJson }),
-})
-
-function mockPluginTempFolder(indexJs: string, pluginJson?: string): [Plugin, () => void] {
-    const folder = fs.mkdtempSync(path.join(os.tmpdir(), 'foo-'))
-
-    fs.writeFileSync(path.join(folder, 'index.js'), indexJs)
-    fs.writeFileSync(
-        path.join(folder, 'plugin.json'),
-        pluginJson ||
-            JSON.stringify({
-                name: 'posthog-maxmind-plugin',
-                description: 'just for testing',
-                url: 'http://example.com/plugin',
-                config: {},
-                main: 'index.js',
-            })
-    )
-    return [
-        { ...plugin60, url: `file:${folder}`, archive: null },
-        () => {
-            fs.rmdirSync(folder, { recursive: true })
-        },
-    ]
-}
-
-const plugin60: Plugin = {
-    id: 60,
-    name: 'posthog-maxmind-plugin',
-    description: 'Ingest GeoIP data via MaxMind',
-    url: 'https://www.npmjs.com/package/posthog-maxmind-plugin',
-    config_schema: {
-        localhostIP: {
-            hint: 'Useful if testing locally',
-            name: 'IP to use instead of 127.0.0.1',
-            type: 'string',
-            order: 2,
-            default: '',
-            required: false,
-        },
-        maxmindMmdb: {
-            hint: 'The "GeoIP2 City" or "GeoLite2 City" database file',
-            name: 'GeoIP .mddb database',
-            type: 'attachment',
-            order: 1,
-            markdown:
-                'Sign up for a [MaxMind.com](https://www.maxmind.com) account, download and extract the database and then upload the `.mmdb` file below',
-            required: true,
-        },
-    },
-    tag: '0.0.2',
-    archive: createZipBuffer('posthog-maxmind-plugin', {
-        indexJs:
-            'function processEvent (event) { if (event.properties) { event.properties.processed = true } return event }',
-    }),
-    error: undefined,
-}
-
-const pluginAttachment1: PluginAttachmentDB = {
-    id: 1,
-    key: 'maxmindMmdb',
-    content_type: 'application/octet-stream',
-    file_name: 'test.txt',
-    file_size: 4,
-    contents: Buffer.from('test'),
-    plugin_config_id: 39,
-    team_id: 2,
-}
-
-const pluginConfig39: PluginConfig = {
-    id: 39,
-    team_id: 2,
-    plugin_id: 60,
-    enabled: true,
-    order: 0,
-    config: { localhostIP: '94.224.212.175' },
-    error: undefined,
-}
 
 let mockServer: PluginsServer
 beforeEach(async () => {
@@ -299,7 +186,7 @@ test('archive plugin with broken plugin.json does not do much', async () => {
 
     expect(setError).toHaveBeenCalled()
     expect(setError.mock.calls[0][0]).toEqual(mockServer)
-    expect(setError.mock.calls[0][1]!.message).toEqual('Can not load plugin.json for plugin "posthog-maxmind-plugin"')
+    expect(setError.mock.calls[0][1]!.message).toEqual('Can not load plugin.json for plugin "test-maxmind-plugin"')
     expect(setError.mock.calls[0][1]!.time).toBeDefined()
     expect(pluginConfigs.get(39)!.vm).toEqual(null)
 })
